+++ packages/playwright-core/src/server/injected/highlight.ts
@@ -45,8 +45,9 @@ export class Highlight {
     this._actionPointElement = document.createElement('x-pw-action-point');
     this._actionPointElement.setAttribute('hidden', 'true');
 
-    // Use a closed shadow root to prevent selectors matching our internal previews.
-    this._glassPaneShadow = this._glassPaneElement.attachShadow({ mode: isUnderTest ? 'open' : 'closed' });
+    // NOTE: do not use an open shadow root, event for test.
+    // Closed shadow root prevents selectors matching our internal previews.
+    this._glassPaneShadow = this._glassPaneElement.attachShadow({ mode: 'closed' });
     this._glassPaneShadow.appendChild(this._actionPointElement);
     const styleElement = document.createElement('style');
     styleElement.textContent = `
@@ -158,6 +159,8 @@ export class Highlight {
         tooltipElement.style.top = '0';
         tooltipElement.style.left = '0';
         tooltipElement.style.display = 'flex';
+        if (this._isUnderTest)
+          console.error('Highlight text for test: ' + JSON.stringify(tooltipElement.textContent)); // eslint-disable-line no-console
       }
       this._highlightEntries.push({ targetElement: elements[i], tooltipElement, highlightElement });
     }
